TableManager : vérifie que le fichier existe avant de le supprimer.
Si une table a été supprimée manuellement (suppression du fichier dans le répertoire uploads/tables), on ne pouvait pas supprimer la table de la liste des tables, car le unlink générait une erreur.

+++ class/Table/TableManager.php
@@ -426,9 +426,11 @@ class TableManager {
         unset($this->opened[$name]);
 
         // Supprime le fichier CSV
-        if (! @unlink($path)) {
-            $msg = __('Impossible de supprimer la table "%s".', 'docalist-core');
-            throw new Exception(sprintf($msg, $name));
+        if (file_exists($path)) {
+            if (! @unlink($path)) {
+                $msg = __('Impossible de supprimer la table "%s".', 'docalist-core');
+                throw new Exception(sprintf($msg, $name));
+            }
         }
 
         // Supprime l'ancienne table du cache
